+++ agent/agent.go
@@ -1,180 +1,31 @@
 package agent
 
 import (
-	"fmt"
-	"time"
-
 	"github.com/pathfinder-cm/pathfinder-agent/daemon"
-	"github.com/pathfinder-cm/pathfinder-agent/metrics"
-	"github.com/pathfinder-cm/pathfinder-agent/util"
 	"github.com/pathfinder-cm/pathfinder-go-client/pfclient"
-	"github.com/pathfinder-cm/pathfinder-go-client/pfmodel"
-	log "github.com/sirupsen/logrus"
 )
 
 type Agent interface {
-	Run(agentType string)
-	RunMetrics() error
-	Process() bool
-	provisionContainer(sc pfmodel.Container, lcs *pfmodel.ContainerList) (bool, error)
-}
-
-type agent struct {
-	nodeHostname    string
-	containerDaemon daemon.ContainerDaemon
-	pfclient        pfclient.Pfclient
+	Run()
 }
 
 func NewAgent(
 	nodeHostname string,
 	containerDaemon daemon.ContainerDaemon,
-	pfclient pfclient.Pfclient) Agent {
+	pfclient pfclient.Pfclient,
+	agentType string) Agent {
 
-	return &agent{
-		nodeHostname:    nodeHostname,
-		containerDaemon: containerDaemon,
-		pfclient:        pfclient,
-	}
-}
-
-func (a *agent) Run(agentType string) {
-	if agentType == "metrics" {
-		log.WithFields(log.Fields{}).Warn("Push Metrics")
-
-		for {
-			delay := 60 + util.RandomIntRange(1, 10)
-			time.Sleep(time.Duration(delay) * time.Second)
-
-			a.RunMetrics()
+	if agentType == "provision" {
+		return &provisionAgent{
+			nodeHostname:    nodeHostname,
+			containerDaemon: containerDaemon,
+			pfclient:        pfclient,
 		}
-		return
-	}
-
-	for {
-		// Add delay between processing
-		delay := 5 + util.RandomIntRange(1, 5)
-		time.Sleep(time.Duration(delay) * time.Second)
-
-		a.Process()
-	}
-}
-
-func (a *agent) RunMetrics() error {
-	m := metrics.NewMetrics()
-	collectedMetrics := m.Collect()
-	err := a.pfclient.PushMetrics(collectedMetrics)
-	if err != nil {
-		log.Error(err.Error())
-		return err
-	}
-
-	return nil
-}
-
-func (a *agent) Process() bool {
-	scs, err := a.pfclient.FetchContainersFromServer(a.nodeHostname)
-	if err != nil {
-		return false
-	}
-
-	// Compare containers between server and local daemon
-	// Do action as necessary
-	for _, sc := range *scs {
-		lcs, err := a.containerDaemon.ListContainers()
-		if err != nil {
-			return false
-		}
-
-		switch status := sc.Status; status {
-		case "SCHEDULED":
-			a.provisionContainer(sc, lcs)
-		case "SCHEDULE_DELETION":
-			a.deleteContainer(sc, lcs)
-		}
-	}
-
-	//TODO: get a list of orphaned containers
-
-	return true
-}
-
-func (a *agent) provisionContainer(sc pfmodel.Container, lcs *pfmodel.ContainerList) (bool, error) {
-	i := lcs.FindByHostname(sc.Hostname)
-	if i == -1 {
-		log.WithFields(log.Fields{
-			"hostname": sc.Hostname,
-			"image":    sc.Image,
-		}).Info("Creating container")
-
-		ok, ipaddress, err := a.containerDaemon.CreateContainer(sc.Hostname, sc.Image)
-		if !ok {
-			a.pfclient.MarkContainerAsProvisionError(
-				a.nodeHostname,
-				sc.Hostname,
-			)
-			log.WithFields(log.Fields{
-				"hostname": sc.Hostname,
-				"image":    sc.Image,
-			}).Error(fmt.Sprintf("Error during container creation. %v", err))
-			return false, err
-		}
-
-		a.pfclient.UpdateIpaddress(
-			a.nodeHostname,
-			sc.Hostname,
-			ipaddress,
-		)
-		a.pfclient.MarkContainerAsProvisioned(
-			a.nodeHostname,
-			sc.Hostname,
-		)
-
-		log.WithFields(log.Fields{
-			"hostname": sc.Hostname,
-			"image":    sc.Image,
-		}).Info("Container created")
 	} else {
-		log.WithFields(log.Fields{
-			"hostname": sc.Hostname,
-			"image":    sc.Image,
-		}).Info("Container already exist")
-
-		a.pfclient.MarkContainerAsProvisioned(
-			a.nodeHostname,
-			sc.Hostname,
-		)
-	}
-
-	return true, nil
-}
-
-func (a *agent) deleteContainer(sc pfmodel.Container, lcs *pfmodel.ContainerList) (bool, error) {
-	i := lcs.FindByHostname(sc.Hostname)
-	if i == -1 {
-		log.WithFields(log.Fields{
-			"hostname": sc.Hostname,
-			"image":    sc.Image,
-		}).Info("Container already deleted")
-	} else {
-		ok, err := a.containerDaemon.DeleteContainer(sc.Hostname)
-		if !ok {
-			log.WithFields(log.Fields{
-				"hostname": sc.Hostname,
-				"image":    sc.Image,
-			}).Error("Error during container deletion")
-			return false, err
+		return &metricsAgent{
+			nodeHostname:    nodeHostname,
+			containerDaemon: containerDaemon,
+			pfclient:        pfclient,
 		}
-
-		log.WithFields(log.Fields{
-			"hostname": sc.Hostname,
-			"image":    sc.Image,
-		}).Info("Container deleted")
 	}
-
-	a.pfclient.MarkContainerAsDeleted(
-		a.nodeHostname,
-		sc.Hostname,
-	)
-
-	return true, nil
 }