+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 #### 0.3.0
 - Add feature to store metrics from node(s) such as memory info (only this for now)
+
 #### 0.2.0
 - Model and pfclient are extracted to [pathfinder-go-client](https://github.com/pathfinder-cm/pathfinder-go-client) repo
+
 #### 0.1.0
 - Initial release: base agent functionality up and running.